Implement iframesTimeout option
In case the load event of an iframe isn't called – e.g. if there's no network connection or the user browses inside the "offline mode" and an iframe has an online "src" – mark.js wasn't able to complete the process for the rest of the site. The now implemented option `iframesTimeout` specifies the maximum ms to wait for a load event. If it's not fired in this time, the iframe will be silently skipped.